Add files via upload
India has the world’s largest education system. Outside the more selective part of private schools, there are no proper facilities, and attention given to the children in rural areas.
In a country with more than 1 billion people with 65% of population in the working age group, there is still  a drop of students when they reach higher classes. A country that gave birth to famous scientists, mathematicians, and business tycoon, like Satya Nadella (Microsoft, CEO), Sundar Pichai (Google, CEO), Shantanu Narayen (Adobe, CEO), etc. is facing this educational crisis.
Therefore, through this analysis, the aim is to form a comparative study in the field of Educational Data Mining field where Primary schools, Upper Primary with Secondary schools are being targeted with the help of Data Mining techniques. Furthermore, with the help of a Decision Tree’s CART Machine Learning Algorithm, a model to predict the total literacy rate of various states of India will be presented.